@@ -1,27 +1,28 @@
 module humungus.tedunangst.com/r/honk
 
-go 1.18
+go 1.20
 
 require (
 	github.com/gorilla/mux v1.8.1
-	github.com/mattn/go-runewidth v0.0.15
-	github.com/mmcdole/gofeed v1.2.1
-	golang.org/x/crypto v0.19.0
-	golang.org/x/net v0.21.0
+	github.com/mattn/go-runewidth v0.0.16
+	github.com/mmcdole/gofeed v1.3.0
+	golang.org/x/crypto v0.26.0
+	golang.org/x/net v0.28.0
 	humungus.tedunangst.com/r/go-sqlite3 v1.2.1
 	humungus.tedunangst.com/r/gonix v0.1.4
-	humungus.tedunangst.com/r/webs v0.7.11
+	humungus.tedunangst.com/r/termvc v0.1.3
+	humungus.tedunangst.com/r/webs v0.7.18
 )
 
 require (
-	github.com/PuerkitoBio/goquery v1.8.0 // indirect
-	github.com/andybalholm/cascadia v1.3.1 // indirect
+	github.com/PuerkitoBio/goquery v1.9.2 // indirect
+	github.com/andybalholm/cascadia v1.3.2 // indirect
 	github.com/json-iterator/go v1.1.12 // indirect
-	github.com/mmcdole/goxpp v1.1.0 // indirect
+	github.com/mmcdole/goxpp v1.1.1 // indirect
 	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
 	github.com/modern-go/reflect2 v1.0.2 // indirect
 	github.com/rivo/uniseg v0.4.7 // indirect
-	golang.org/x/image v0.15.0 // indirect
-	golang.org/x/sys v0.17.0 // indirect
-	golang.org/x/text v0.14.0 // indirect
+	golang.org/x/image v0.19.0 // indirect
+	golang.org/x/sys v0.23.0 // indirect
+	golang.org/x/text v0.17.0 // indirect
 )
