@@ -194,7 +194,7 @@ function statechanger(evt) {
 	}
 	switchtopage(data.name, data.arg)
 }
-function switchtopage(name, arg) {
+function switchtopage(name, arg, anchor) {
 	var stash = curpagestate.name + ":" + curpagestate.arg
 	var honksonpage = document.getElementById("honksonpage")
 	var holder = honksonpage.children[0]
@@ -220,6 +220,10 @@ function switchtopage(name, arg) {
 		if (msg) {
 			srvel.prepend(msg)
 		}
+		if (anchor) {
+			let el = document.getElementById(anchor)
+			el.scrollIntoView()
+		}
 	} else {
 		// or create one and fill it
 		honksonpage.prepend(document.createElement("div"))
@@ -227,6 +231,10 @@ function switchtopage(name, arg) {
 		get("/hydra?" + encode(args), function(xhr) {
 			if (xhr.status == 200) {
 				fillinhonks(xhr, false)
+				if (anchor) {
+					let el = document.getElementById(anchor)
+					el.scrollIntoView()
+				}
 			} else {
 				refreshupdate(" status: " + xhr.status)
 			}
@@ -246,11 +254,14 @@ function pageswitcher(name, arg) {
 		if (name == curpagestate.name && arg == curpagestate.arg) {
 			return false
 		}
-		switchtopage(name, arg)
-		var url = evt.srcElement.href
-		if (!url) {
+		let url = evt.srcElement.href
+		if (!url)
 			url = evt.srcElement.parentElement.href
-		}
+		let anchor
+		let arr = url.split("#")
+		if (arr.length == 2)
+			anchor = arr[1]
+		switchtopage(name, arg, anchor)
 		history.pushState(newpagestate(name, arg), "some title", url)
 		window.scrollTo(0, 0)
 		return false
@@ -377,12 +388,35 @@ function showhonkform(elem, rid, hname) {
 		ridinput.value = ""
 		honknoise.value = ""
 	}
+	honknoise.ondrop = donkdrop
 	var updateinput = document.getElementById("updatexidinput")
 	updateinput.value = ""
 	var savedfile = document.getElementById("saveddonkxid")
 	savedfile.value = ""
 	honknoise.focus()
 	return false
+}
+function donkdrop(evt) {
+	evt.preventDefault()
+	let donks = document.querySelector("#donker input")
+	Array.from(evt.dataTransfer.items).forEach((item) => {
+        if (item.kind == "file") {
+			let olddonks = donks.files
+			let donkarama = new DataTransfer();
+			for (donk of olddonks)
+				donkarama.items.add(donk)
+            let file = item.getAsFile()
+			donkarama.items.add(file)
+			donks.files = donkarama.files;
+            let t = evt.target
+            let start = t.selectionStart
+            let s = t.value.substr(0, start)
+            let e = t.value.substr(start)
+            t.value = s + `<img src=${donks.files.length}>` + e
+            t.selectionStart = start
+            t.selectionEnd = start
+        }
+    })
 }
 function cancelhonking() {
 	hideelement(lehonkform)
